+++ src/index.ts
@@ -3,5 +3,6 @@ export * from "./scalars";
 
 export { buildSchema, BuildSchemaOptions } from "./utils/buildSchema";
 export { useContainer } from "./utils/container";
+export { ResolverInterface } from "./interfaces/ResolverInterface";
 export { ArgumentValidationError } from "./errors/ArgumentValidationError";
 export { formatArgumentValidationError } from "./errors/formatArgumentValidationError";