+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 ### Features
 - add basic support for automatic arguments and inputs validation using `class-validator`
+- add interface `ResolverInterface` for type checking of resolver class methods (field resolvers)
 
 ## v0.3.0
 ### Features